Consolidate progression and fetch helpers
