+++ Javascript/resourceBar.js
@@ -3,7 +3,7 @@
 // Version:  7/1/2025 10:38
 // Developer: Deathsgift66
 import { supabase } from '../supabaseClient.js';
-import { fetchJson } from './fetchJson.js'; // Custom wrapper for JSON fetch
+import { fetchJson } from './utils.js'; // Custom wrapper for JSON fetch
 
 // Dynamically load Bootstrap 5.3 if not already loaded
 function injectBootstrap(onReady) {
