+++ alliance_quests.html
@@ -43,7 +43,7 @@ Developer: Deathsgift66
 
   <!-- JS Modules -->
   <script src="/Javascript/allianceAppearance.js" type="module"></script>
-  <script src="/Javascript/progressionBanner.js" type="module"></script>
+  <script src="/Javascript/progression.js" type="module"></script>
   <script src="/Javascript/alliance_quests.js" type="module"></script>
 
 <!-- ✅ Injected standard Thronestead modules -->
