+++ alliance_treaties.html
@@ -46,7 +46,7 @@ Developer: Deathsgift66
 
   <!-- JS Modules -->
   <script src="/Javascript/allianceAppearance.js" type="module"></script>
-  <script src="/Javascript/progressionBanner.js" type="module"></script>
+  <script src="/Javascript/progression.js" type="module"></script>
   <script src="/Javascript/alliance_treaties.js" type="module"></script>
 
 <!-- ✅ Injected standard Thronestead modules -->
