+++ alliance_wars.html
@@ -43,7 +43,7 @@ Developer: Deathsgift66
 
   <!-- Scripts -->
   <script src="/Javascript/allianceAppearance.js" type="module"></script>
-  <script src="/Javascript/progressionBanner.js" type="module"></script>
+  <script src="/Javascript/progression.js" type="module"></script>
   <script src="/Javascript/alliance_wars.js" type="module"></script>
 
   <!-- ✅ Injected standard Thronestead modules -->
