+++ compose.html
@@ -34,7 +34,7 @@ Developer: Deathsgift66
   <link href="/CSS/compose.css" rel="stylesheet" />
   <script type="module">
     import { supabase } from '/Javascript/supabaseClient.js';
-    import { authFetchJson } from '/Javascript/fetchJson.js';
+    import { authFetchJson } from '/Javascript/utils.js';
     import { showToast, getValue, setText, escapeHTML } from '/Javascript/utils.js';
 
     let session;
