+++ diplomacy_center.html
@@ -39,7 +39,7 @@ Developer: Deathsgift66
   <link href="/CSS/diplomacy_center.css" rel="stylesheet" />
 
   <!-- Scripts -->
-  <script src="/Javascript/progressionBanner.js" type="module"></script>
+  <script src="/Javascript/progression.js" type="module"></script>
 
 <!-- ✅ Injected standard Thronestead modules -->
   <script src="/Javascript/components/authGuard.js" type="module"></script>
