+++ kingdom_list.html
@@ -14,7 +14,7 @@
   <script src="/Javascript/apiHelper.js" type="module"></script>
   <script src="/Javascript/navLoader.js" type="module"></script>
   <script type="module">
-    import { fetchJson } from './Javascript/fetchJson.js';
+    import { fetchJson } from './Javascript/utils.js';
     document.addEventListener('DOMContentLoaded', async () => {
       try {
         const kingdoms = await fetchJson('/api/kingdom/lookup');
