+++ kingdom_profile.html
@@ -46,7 +46,7 @@ Developer: Deathsgift66
     // Version:  7/1/2025 10:38
     // Developer: Deathsgift66
     import { supabase } from '/Javascript/supabaseClient.js';
-    import { authFetchJson } from './fetchJson.js';
+    import { authFetchJson } from '/Javascript/utils.js';
     import { openModal, closeModal } from '/Javascript/utils.js';
 
     let targetKingdomId = null;
