+++ policies_laws.html
@@ -247,7 +247,7 @@ function updateSummary(activePolicyId, activeLawsIds, policies, laws) {
   <link href="/CSS/kr_navbar.css" rel="stylesheet" />
   <link href="/CSS/progressionBanner.css" rel="stylesheet" />
   <link href="/CSS/resource_bar.css" rel="stylesheet" />
-  <script src="/Javascript/progressionBanner.js" type="module"></script>
+  <script src="/Javascript/progression.js" type="module"></script>
 
 <!-- ✅ Injected standard Thronestead modules -->
   <script src="/Javascript/components/authGuard.js" type="module"></script>
