+++ resources.html
@@ -116,7 +116,7 @@ Developer: Deathsgift66
 // Version:  7/1/2025 10:38
 // Developer: Deathsgift66
 import { supabase } from '/Javascript/supabaseClient.js';
-import { fetchJson } from './fetchJson.js';
+import { fetchJson } from './Javascript/utils.js';
 
 let krSub = null,
   vaultSub = null,
