+++ villages.html
@@ -38,7 +38,7 @@ Developer: Deathsgift66
   <link href="/CSS/root_theme.css" rel="stylesheet" />
   <link href="/CSS/kr_navbar.css" rel="stylesheet" />
   <link href="/CSS/progressionBanner.css" rel="stylesheet" />
-  <script src="/Javascript/progressionBanner.js" type="module"></script>
+  <script src="/Javascript/progression.js" type="module"></script>
   <link href="/CSS/resource_bar.css" rel="stylesheet" />
 
 <!-- ✅ Injected standard Thronestead modules -->
